requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Werkzeug==2.0.1
Jinja2==3.0.1
SQLAlchemy==1.4.23
gunicorn==20.1.0
python-dotenv==0.19.0
stripe==4.2.0
Flask-Migrate==3.1.0 